+++ src/mod.ts
@@ -1,1 +1,3 @@
-export { StellarAssetsSdk } from "./stellar-assets-sdk.ts";
+export { Sdk } from "./sdk.ts";
+export * from "./types.ts";
+export * from "./utils.ts";